slash cmds wip 001
